Address PR #790 review finding: format fan tooltip value as float, not unsigned int
fanSpeedPercent is intentionally left unclamped, so a corrupted-input reading
can exceed unsigned int's range; static_cast<unsigned int> on an out-of-range
float is undefined behavior. Format the float value directly instead
(matching how the NaN case right above it is already handled specially).

+++ src/App/Panels/GpuSection.cpp
@@ -598,7 +598,11 @@ void renderGpuSection(RenderContext& ctx)
                                     }
                                     else
                                     {
-                                        ImGui::TextColored(theme.scheme().gpuFan, "Fan: %u%%", static_cast<unsigned int>(fanData[fanIdx]));
+                                        // Format the float directly rather than narrowing to unsigned int: this PR
+                                        // intentionally leaves fanSpeedPercent unclamped (see GPUModel::computeSnapshot),
+                                        // so a corrupted-input value can exceed unsigned int's range, and casting an
+                                        // out-of-range float to an integer type is undefined behavior.
+                                        ImGui::TextColored(theme.scheme().gpuFan, "Fan: %.0f%%", static_cast<double>(fanData[fanIdx]));
                                     }
                                 }
                             }
